requirements.txt
streamlit==1.40.1
langchain== 0.3.21
langchain_huggingface==0.1.2
fuzzywuzzy==0.18.0
python-dotenv==0.21.0
langchain_pinecone==0.2.3
langchain_openai==0.3.10
sentence-transformers==3.4.1
